pylbo: refactored mode figures for easier handling
